S9/cloud_function_storage/requirements.txt
Flask==3.1.2
functions-framework==3.10.0
google-cloud-storage>=2.10.0
